IMS: Dialer app changes for IMS
IMS Conference URI UI changes
1. Add a new button for initiating conference URI
call in the Dial screen.
2. Send a intent for the conference call with a
extra to indicate the call type as conference
CRs-Fixed: 455241

Enable visibility of digits container for adding call
1. When a call is active, the user goes to home screen and presses
the phone button and then chooses add call. The recipients box
is not visible as the visibility of digits container is not
VISIBLE. In this change, the visibility of digits container is
set to VISBILE
CRs-Fixed: 486790

Change-Id: Id0131d5063e2c95647a6f956d0d1a39128fba940
CRs-Fixed: 582011